Apply enroute survival to adult salmon

This function calculates the survival of adult salmon during their enroute migration.

apply_enroute_survival(
  year,
  month,
  adults,
  month_return_proportions,
  gates_overtopped,
  tisdale_bypass_watershed,
  yolo_bypass_watershed,
  migratory_temperature_proportion_over_20,
  ..surv_adult_enroute_int,
  .adult_en_route_migratory_temp,
  .adult_en_route_bypass_overtopped,
  hatchery_release,
  stochastic
)

Arguments

year

The year of the simulation.

month

The starting month of the simulation.

adults

A data frame containing information about adult salmon.

month_return_proportions

A vector of proportions for each month.

gates_overtopped

A three-dimensional array containing information about gates overtopped.

tisdale_bypass_watershed

A parameter related to the Tisdale bypass watershed.

yolo_bypass_watershed

A parameter related to the Yolo bypass watershed.

migratory_temperature_proportion_over_20

A matrix containing proportions of migratory temperatures over 20 degrees.

..surv_adult_enroute_int

Additional parameters for adult enroute survival.

.adult_en_route_migratory_temp

Parameter for adult enroute survival related to migratory temperature.

.adult_en_route_bypass_overtopped

Parameter for adult enroute survival related to bypass overtopped.

stochastic

Logical indicating whether to include stochastic elements in the simulation.

.adult_en_route_adult_harvest_rate

Parameter for adult enroute survival related to adult harvest rate.

Value

A list containing the initial number of adults, the proportion of natural adults, and the initial number of adults by month.