Get habitat Values

Returns the appropriate habitat for spawning and rearing based on habitat inputs and a given simulation year and month

get_habitat(
  year,
  month,
  inchannel_habitat_fry,
  inchannel_habitat_juvenile,
  floodplain_habitat,
  sutter_habitat,
  yolo_habitat,
  delta_habitat
)

Source

IP-117068

Arguments

year

The current simulation year, 1-20

month

The current simulation month, 1-8

inchannel_habitat_fry

3 dimensional array [watersheds, months, years] representing fry inchannel habitat in square meters

inchannel_habitat_juvenile

3 dimensional array [watersheds, months, years] representing juvenile inchannel habitat in square meters

floodplain_habitat

3 dimensional array [watersheds, months, years] representing floodplain habitat in square meters

sutter_habitat

2 dimensional array [months, years] representing sutter bypass habitat in square meters

yolo_habitat

2 dimensional array [months, years] representing yolo bypass habitat in square meters

north_delta_habitat

2 dimensional array [months, years] representing north delta habitat in square meters

south_delta_habitat

2 dimensional array [months, years] representing south delta habitat in square meters

Details

See params for details on parameter sources