Get Migratory Survival Rates

Calculates the juvenile out migration survival rates in all regions for a month and year of the simulation

get_migratory_survival(
  year,
  month,
  cc_gates_prop_days_closed,
  freeport_flows,
  vernalis_flows,
  stockton_flows,
  vernalis_temps,
  prisoners_point_temps,
  CVP_exports,
  SWP_exports,
  upper_sacramento_flows,
  delta_inflow,
  avg_temp_delta,
  avg_temp,
  delta_proportion_diverted,
  .surv_juv_outmigration_sac_delta_intercept_one,
  .surv_juv_outmigration_sac_delta_intercept_two,
  .surv_juv_outmigration_sac_delta_intercept_three,
  .surv_juv_outmigration_sac_delta_delta_flow,
  .surv_juv_outmigration_sac_delta_avg_temp,
  .surv_juv_outmigration_sac_delta_perc_diversions,
  .surv_juv_outmigration_sac_delta_medium,
  .surv_juv_outmigration_sac_delta_large,
  ..surv_juv_outmigration_sj_int,
  ..surv_juv_outmigration_sac_int_one,
  ..surv_juv_outmigration_sac_prop_diversions,
  ..surv_juv_outmigration_sac_total_diversions,
  ..surv_juv_outmigration_sac_int_two,
  .surv_juv_outmigration_san_joaquin_medium,
  .surv_juv_outmigration_san_joaquin_large,
  min_survival_rate,
  surv_juv_outmigration_sac_delta_model_weights
)

Arguments

year
The simulation year, 1-20

month
The simulation month, 1-8

cc_gates_prop_days_closed
More details at delta_cross_channel_closed

freeport_flows
More details at freeport_flow

vernalis_flows
More details at vernalis_flow

stockton_flows
More details at stockton_flow

vernalis_temps
More details at vernalis_temperature

prisoners_point_temps
More details at prisoners_point_temperature

CVP_exports
More details at cvp_exports

SWP_exports
More details at swp_exports

.surv_juv_outmigration_sac_delta_intercept_one
Intercept surv_juv_outmigration_sac_delta for model one

.surv_juv_outmigration_sac_delta_intercept_two
Intercept surv_juv_outmigration_sac_delta for model two

.surv_juv_outmigration_sac_delta_intercept_three
Intercept surv_juv_outmigration_sac_delta for model three

.surv_juv_outmigration_sac_delta_delta_flow
Coefficient surv_juv_outmigration_sac_delta for delta_flow variable

.surv_juv_outmigration_sac_delta_avg_temp
Coefficient surv_juv_outmigration_sac_delta for avg_temp variable

.surv_juv_outmigration_sac_delta_perc_diversions
Coefficient surv_juv_outmigration_sac_delta for perc_diversions variable

.surv_juv_outmigration_sac_delta_medium
Size related intercept for surv_juv_outmigration_sac_delta medium sized fish

.surv_juv_outmigration_sac_delta_large
Size related intercept for surv_juv_outmigration_sac_delta large sized fish

..surv_juv_outmigration_sj_int
Intercept for surv_juv_outmigration_san_joaquin

.surv_juv_outmigration_san_joaquin_medium
Size related intercept for surv_juv_outmigration_san_joaquin medium sized fish

.surv_juv_outmigration_san_joaquin_large
Size related intercept for surv_juv_outmigration_san_joaquin large sized fish

min_survival_rate
estimated survival rate if temperature threshold is exceeded

surv_juv_outmigration_sac_delta_model_weights
weights for competing models

Source

IP-117068