Harvest Adults Function

This function calculates the total adult salmon harvest based on various parameters. This function is called within the larger model.R file before stray rates and prespawn survival is applied.

harvest_adults(
  adult_df,
  spawner_df,
  year = year,
  spawn_habitat,
  terminal_hatchery_logic = c(T, F),
  ocean_harvest_percentage,
  tributary_harvest_percentage,
  restrict_harvest_to_hatchery_ocean = c(F, T),
  restrict_harvest_to_hatchery_trib = c(F, T),
  no_cohort_harvest_years = NULL,
  intelligent_habitat_harvest = c(F, T),
  intelligent_crr_harvest = c(F, T),
  crr_scaling = 2
)

Arguments

adult_df

A data frame containing information about adult salmon.

year

The target year for the harvest calculation.

ocean_harvest_percentage

The percentage of harvest from the ocean, same for all tributaries.

tributary_harvest_percentage

The percentage of harvest from tributaries, can vary by tributary.

no_cohort_harvest_years

A vector of years where no cohort harvest is allowed.

intelligent_crr_harvest

A logical vector indicating whether to apply intelligent harvest based on Coho Retention Rate (CRR).

restrict_harvest_to_hatchery

A logical vector indicating whether to restrict harvest to hatcheries for each tributary.

Value

A numeric value representing the total adult salmon harvest.