Juvenile Month Dynamics

Function to run through monthly juvenile routing and rearing logic.

juvenile_month_dynamic(
  fish,
  year = year,
  month = month,
  rearing_survival = rearing_survival,
  migratory_survival = migratory_survival,
  habitat = habitat,
  ..params = ..params,
  avg_ocean_transition_month = avg_ocean_transition_month,
  stochastic = stochastic,
  juveniles,
  ic_growth,
  fp_growth,
  delta_growth,
  filling_fn = fallRunDSM::fill_natal,
  filling_args = NULL,
  filling_regional_fn = fallRunDSM::fill_regional,
  filling_regional_args = NULL,
  movement_fn = NULL,
  movement_args = NULL,
  movement_months = NULL
)

Arguments

fish

list object containing tracking matrices for juveniles, juveniles at chips, and adults in ocean

year

year

month

month

rearing_survival

rearing survival for year and month

migratory_survival

migratory survival for year and month

habitat

habitat for year and month

..params

full params list from model parms

hypothesis

movement hypothesis to test

Examples

fish = list(juveniles = juveniles,
            north_delta_fish = north_delta_fish,
            south_delta_fish = south_delta_fish,
            juveniles_at_chipps = juveniles_at_chipps,
            adults_in_ocean = adults_in_ocean)
#> Error: object 'juveniles' not found