Juvenile Rearing Survival

Calculates the juvenile rearing survival inchannel and on the floodplain

surv_juv_rear(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diversions,
  total_diversions,
  stranded,
  weeks_flooded,
  ..surv_juv_rear_int = winterRunDSM::params$..surv_juv_rear_int,
  .avg_temp_thresh = winterRunDSM::params$.surv_juv_rear_avg_temp_thresh,
  .high_predation = winterRunDSM::params$.surv_juv_rear_high_predation,
  .surv_juv_rear_contact_points = winterRunDSM::params$.surv_juv_rear_contact_points,
  ..surv_juv_rear_contact_points = winterRunDSM::params$..surv_juv_rear_contact_points,
  .surv_juv_rear_prop_diversions = winterRunDSM::params$.surv_juv_rear_prop_diversions,
 
    ..surv_juv_rear_prop_diversions = winterRunDSM::params$..surv_juv_rear_prop_diversions,
 
    .surv_juv_rear_total_diversions = winterRunDSM::params$.surv_juv_rear_total_diversions,
 
    ..surv_juv_rear_total_diversions = winterRunDSM::params$..surv_juv_rear_total_diversions,
  .stranded = winterRunDSM::params$.surv_juv_rear_stranded,
  .medium = winterRunDSM::params$.surv_juv_rear_medium,
  .large = winterRunDSM::params$.surv_juv_rear_large,
  .floodplain = winterRunDSM::params$.surv_juv_rear_floodplain,
  min_survival_rate = winterRunDSM::params$min_survival_rate
)

Arguments

max_temp_thresh
variable representing probability of exceeding the max temperature threshold

avg_temp_thresh
variable representing probability of exceeding the avg temperature threshold

high_predation
variable representing indicator of high predation for a watershed

contact_points
variable representing total number of contact points per watershed

prop_diversions
variable representing proportion of water diverted

total_diversions
variable representing total amount of water diverted

stranded
variable representing stranding rate per watershed

weeks_flooded
variable representing total weeks flooded per watershed

..surv_juv_rear_int
intercept, source: calibration (varies by tributary)

.avg_temp_thresh
coefficient for avg_temp_thresh variable

.high_predation
coefficient for high_predation variable

.surv_juv_rear_contact_points
coefficient for contact_points variable

..surv_juv_rear_contact_points
calibrated coefficient for contact_points variable

.surv_juv_rear_prop_diversions
coefficient for prop_diversions variable

..surv_juv_rear_prop_diversions
calibrated coefficient for prop_diversions variable

.surv_juv_rear_total_diversions
coefficient for total_diversions variable

..surv_juv_rear_total_diversions
calibrated coefficient for total_diversions variable

.stranded
coefficient for stranded variable

.medium
size related intercept for medium sized fish

.large
size related intercept for large sized fish

.floodplain
Additional intercept for floodplain rearing benefit

min_survival_rate
estimated survival rate if temperature threshold is exceeded, source: expert opinion

Source

IP-117068

Details

See params for details on parameter sources