Fall Run Chinook Model

Fall Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

winter_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = winterRunDSM::params
)

Arguments

scenario
Model inputs, can be modified to test management actions

mode
The mode to run model in. Can be "seed", "simulate", "calibrate"

seeds
The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params
Parameters for model and submodels

Source

IP-117068

Examples

winter_run_seeds <- winterRunDSM::winter_run_model(mode = "seed")
winterRunDSM::winter_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = winter_run_seeds)

#> $spawners
#>                                 1      2      3      4      5      6      7
#> Upper Sacramento River     395580 263559 293165 144002 314422 126157 175001
#> Antelope Creek                  0      0      0      0      0      0      0
#> Battle Creek                    0      0      1      0      0      0      1
#> Bear Creek                      0      0      0      0      0      0      0
#> Big Chico Creek                 0      0      0      0      0      0      0
#> Butte Creek                     0      0      0      0      0      0      0
#> Clear Creek                     0      0      0      0      0      0      0
#> Cottonwood Creek                0      0      0      0      0      0      0
#> Cow Creek                       0      0      0      0      0      0      0
#> Deer Creek                      0      0      0      0      0      0      0
#> Elder Creek                     0      0      0      0      0      0      0
#> Mill Creek                      0      0      0      0      0      0      0
#> Paynes Creek                    0      0      0      0      0      0      0
#> Stony Creek                     0      0      0      0      0      0      0
#> Thomes Creek                    0      0      0      0      0      0      0
#> Upper-mid Sacramento River      0      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0      0
#> Bear River                      0      0      0      0      0      0      0
#> Feather River                   0      0      0      0      0      0      0
#> Yuba River                      0      0      0      0      0      0      0
#> Lower-mid Sacramento River      0      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0      0
#> American River                  0      0      0      0      0      0      0
#> Lower Sacramento River          0      0      0      0      0      0      0
#> Calaveras River                 0      0      0      0      0      0      0
#> Cosumnes River                  0      0      0      0      0      0      0
#> Mokelumne River                 0      0      0      0      0      0      0
#> Merced River                    0      0      0      0      0      0      0
#> Stanislaus River                0      0      0      0      0      0      0
#> Tuolumne River                  0      0      0      0      0      0      0
#> San Joaquin River               0      0      0      0      0      0      0
#>                                 8     9    10     11     12     13     14
#> Upper Sacramento River     257772 83401 66275 404419 103327 294416 364211
#> Antelope Creek                  0     0     0      0      0      0      0
#> Battle Creek                    0     0     0      0      0      0      0
#> Bear Creek                      0     0     0      0      0      0      0
#> Big Chico Creek                 0     0     0      0      0      0      0
#> Butte Creek                     0     0     0      0      0      0      0
#> Clear Creek                     0     0     0      0      0      0      0
#> Cottonwood Creek                0     0     0      0      0      0      0
#> Cow Creek                       0     0     0      0      0      0      0
#> Deer Creek                      0     0     0      0      0      0      0
#> Elder Creek                     0     0     0      0      0      0      0
#> Mill Creek                      0     0     0      0      0      0      0
#> Paynes Creek                    0     0     0      0      0      0      0
#> Stony Creek                     0     0     0      0      0      0      0
#> Thomes Creek                    0     0     0      0      0      0      0
#> Upper-mid Sacramento River      0     0     0      0      0      0      0
#> Sutter Bypass                   0     0     0      0      0      0      0
#> Bear River                      0     0     0      0      0      0      0
#> Feather River                   0     0     0      0      0      0      0
#> Yuba River                      0     0     0      0      0      0      0
#> Lower-mid Sacramento River      0     0     0      0      0      0      0
#> Yolo Bypass                     0     0     0      0      0      0      0
#> American River                  0     0     0      0      0      0      0
#> Lower Sacramento River          0     0     0      0      0      0      0
#> Calaveras River                 0     0     0      0      0      0      0
#> Cosumnes River                  0     0     0      0      0      0      0
#> Mokelumne River                 0     0     0      0      0      0      0
#> Merced River                    0     0     0      0      0      0      0
#> Stanislaus River                0     0     0      0      0      0      0
#> Tuolumne River                  0     0     0      0      0      0      0
#> San Joaquin River               0     0     0      0      0      0      0
#>                                15    16     17     18     19     20
#> Upper Sacramento River     319727 65093 389072 248037 407299 344267
#> Antelope Creek                  0     0      0      0      0      0
#> Battle Creek                    0     0      0      0    208    395
#> Bear Creek                      0     0      0      0      0      0
#> Big Chico Creek                 0     0      0      0      0      0
#> Butte Creek                     0     0      0      0      0      0
#> Clear Creek                     0     0      0      0      0      0
#> Cottonwood Creek                0     0      0      0      0      0
#> Cow Creek                       0     0      0      0      0      0
#> Deer Creek                      0     0      0      0      0      0
#> Elder Creek                     0     0      0      0      0      0
#> Mill Creek                      0     0      0      0      0      0
#> Paynes Creek                    0     0      0      0      0      0
#> Stony Creek                     0     0      0      0      0      0
#> Thomes Creek                    0     0      0      0      0      0
#> Upper-mid Sacramento River      0     0      0      0      0      0
#> Sutter Bypass                   0     0      0      0      0      0
#> Bear River                      0     0      0      0      0      0
#> Feather River                   0     0      0      0      0      0
#> Yuba River                      0     0      0      0      0      0
#> Lower-mid Sacramento River      0     0      0      0      0      0
#> Yolo Bypass                     0     0      0      0      0      0
#> American River                  0     0      0      0      0      0
#> Lower Sacramento River          0     0      0      0      0      0
#> Calaveras River                 0     0      0      0      0      0
#> Cosumnes River                  0     0      0      0      0      0
#> Mokelumne River                 0     0      0      0      0      0
#> Merced River                    0     0      0      0      0      0
#> Stanislaus River                0     0      0      0      0      0
#> Tuolumne River                  0     0      0      0      0      0
#> San Joaquin River               0     0      0      0      0      0
#> 
#> $natural_spawners
#>                               1    2   3   4   5    6    7    8   9 10 11 12 13
#> Upper Sacramento River     1818 1573 713 280 527 1859 3488 2654 749 84 73 23 10
#> Antelope Creek                0    0   0   0   0    0    0    0   0  0  0  0  0
#> Battle Creek                  0    0   1   0   0    0    1    0   0  0  0  0  0
#> Bear Creek                    0    0   0   0   0    0    0    0   0  0  0  0  0
#> Big Chico Creek               0    0   0   0   0    0    0    0   0  0  0  0  0
#> Butte Creek                   0    0   0   0   0    0    0    0   0  0  0  0  0
#> Clear Creek                   0    0   0   0   0    0    0    0   0  0  0  0  0
#> Cottonwood Creek              0    0   0   0   0    0    0    0   0  0  0  0  0
#> Cow Creek                     0    0   0   0   0    0    0    0   0  0  0  0  0
#> Deer Creek                    0    0   0   0   0    0    0    0   0  0  0  0  0
#> Elder Creek                   0    0   0   0   0    0    0    0   0  0  0  0  0
#> Mill Creek                    0    0   0   0   0    0    0    0   0  0  0  0  0
#> Paynes Creek                  0    0   0   0   0    0    0    0   0  0  0  0  0
#> Stony Creek                   0    0   0   0   0    0    0    0   0  0  0  0  0
#> Thomes Creek                  0    0   0   0   0    0    0    0   0  0  0  0  0
#> Upper-mid Sacramento River    0    0   0   0   0    0    0    0   0  0  0  0  0
#> Sutter Bypass                 0    0   0   0   0    0    0    0   0  0  0  0  0
#> Bear River                    0    0   0   0   0    0    0    0   0  0  0  0  0
#> Feather River                 0    0   0   0   0    0    0    0   0  0  0  0  0
#> Yuba River                    0    0   0   0   0    0    0    0   0  0  0  0  0
#> Lower-mid Sacramento River    0    0   0   0   0    0    0    0   0  0  0  0  0
#> Yolo Bypass                   0    0   0   0   0    0    0    0   0  0  0  0  0
#> American River                0    0   0   0   0    0    0    0   0  0  0  0  0
#> Lower Sacramento River        0    0   0   0   0    0    0    0   0  0  0  0  0
#> Calaveras River               0    0   0   0   0    0    0    0   0  0  0  0  0
#> Cosumnes River                0    0   0   0   0    0    0    0   0  0  0  0  0
#> Mokelumne River               0    0   0   0   0    0    0    0   0  0  0  0  0
#> Merced River                  0    0   0   0   0    0    0    0   0  0  0  0  0
#> Stanislaus River              0    0   0   0   0    0    0    0   0  0  0  0  0
#> Tuolumne River                0    0   0   0   0    0    0    0   0  0  0  0  0
#> San Joaquin River             0    0   0   0   0    0    0    0   0  0  0  0  0
#>                            14 15 16 17  18   19   20
#> Upper Sacramento River     14  4 50 89 199 1552 3087
#> Antelope Creek              0  0  0  0   0    0    0
#> Battle Creek                0  0  0  0   0  208  395
#> Bear Creek                  0  0  0  0   0    0    0
#> Big Chico Creek             0  0  0  0   0    0    0
#> Butte Creek                 0  0  0  0   0    0    0
#> Clear Creek                 0  0  0  0   0    0    0
#> Cottonwood Creek            0  0  0  0   0    0    0
#> Cow Creek                   0  0  0  0   0    0    0
#> Deer Creek                  0  0  0  0   0    0    0
#> Elder Creek                 0  0  0  0   0    0    0
#> Mill Creek                  0  0  0  0   0    0    0
#> Paynes Creek                0  0  0  0   0    0    0
#> Stony Creek                 0  0  0  0   0    0    0
#> Thomes Creek                0  0  0  0   0    0    0
#> Upper-mid Sacramento River  0  0  0  0   0    0    0
#> Sutter Bypass               0  0  0  0   0    0    0
#> Bear River                  0  0  0  0   0    0    0
#> Feather River               0  0  0  0   0    0    0
#> Yuba River                  0  0  0  0   0    0    0
#> Lower-mid Sacramento River  0  0  0  0   0    0    0
#> Yolo Bypass                 0  0  0  0   0    0    0
#> American River              0  0  0  0   0    0    0
#> Lower Sacramento River      0  0  0  0   0    0    0
#> Calaveras River             0  0  0  0   0    0    0
#> Cosumnes River              0  0  0  0   0    0    0
#> Mokelumne River             0  0  0  0   0    0    0
#> Merced River                0  0  0  0   0    0    0
#> Stanislaus River            0  0  0  0   0    0    0
#> Tuolumne River              0  0  0  0   0    0    0
#> San Joaquin River           0  0  0  0   0    0    0
#> 
#> $juvenile_biomass
#>                                   1        2       3         4         5
#> Upper Sacramento River     720598.6 822962.6 1273207 3263307.8 4737726.2
#> Antelope Creek                  0.0      0.0       0       0.0       0.0
#> Battle Creek                   31.4    282.6     314     376.8     439.6
#> Bear Creek                      0.0      0.0       0       0.0       0.0
#> Big Chico Creek                 0.0      0.0       0       0.0       0.0
#> Butte Creek                     0.0      0.0       0       0.0       0.0
#> Clear Creek                     0.0      0.0       0       0.0       0.0
#> Cottonwood Creek                0.0      0.0       0       0.0       0.0
#> Cow Creek                       0.0      0.0       0       0.0       0.0
#> Deer Creek                      0.0      0.0       0       0.0       0.0
#> Elder Creek                     0.0      0.0       0       0.0       0.0
#> Mill Creek                      0.0      0.0       0       0.0       0.0
#> Paynes Creek                    0.0      0.0       0       0.0       0.0
#> Stony Creek                     0.0      0.0       0       0.0       0.0
#> Thomes Creek                    0.0      0.0       0       0.0       0.0
#> Upper-mid Sacramento River      0.0      0.0       0       0.0       0.0
#> Sutter Bypass                   0.0      0.0       0       0.0       0.0
#> Bear River                      0.0      0.0       0       0.0       0.0
#> Feather River                   0.0      0.0       0       0.0       0.0
#> Yuba River                      0.0      0.0       0       0.0       0.0
#> Lower-mid Sacramento River      0.0      0.0       0       0.0       0.0
#> Yolo Bypass                     0.0      0.0       0       0.0       0.0
#> American River                  0.0      0.0       0       0.0       0.0
#> Lower Sacramento River          0.0      0.0       0       0.0       0.0
#> Calaveras River                 0.0      0.0       0       0.0       0.0
#> Cosumnes River                  0.0      0.0       0       0.0       0.0
#> Mokelumne River                 0.0      0.0       0       0.0       0.0
#> Merced River                    0.0      0.0       0       0.0       0.0
#> Stanislaus River                0.0      0.0       0       0.0       0.0
#> Tuolumne River                  0.0      0.0       0       0.0       0.0
#> San Joaquin River               0.0      0.0       0       0.0       0.0
#>                                    6         7         8         9        10
#> Upper Sacramento River     4773428.0 5094053.4 5112924.8 5139803.2 5139991.6
#> Antelope Creek                   0.0       0.0       0.0       0.0       0.0
#> Battle Creek                   439.6     502.4     502.4     502.4     502.4
#> Bear Creek                       0.0       0.0       0.0       0.0       0.0
#> Big Chico Creek                  0.0       0.0       0.0       0.0       0.0
#> Butte Creek                      0.0       0.0       0.0       0.0       0.0
#> Clear Creek                      0.0       0.0       0.0       0.0       0.0
#> Cottonwood Creek                 0.0       0.0       0.0       0.0       0.0
#> Cow Creek                        0.0       0.0       0.0       0.0       0.0
#> Deer Creek                       0.0       0.0       0.0       0.0       0.0
#> Elder Creek                      0.0       0.0       0.0       0.0       0.0
#> Mill Creek                       0.0       0.0       0.0       0.0       0.0
#> Paynes Creek                     0.0       0.0       0.0       0.0       0.0
#> Stony Creek                      0.0       0.0       0.0       0.0       0.0
#> Thomes Creek                     0.0       0.0       0.0       0.0       0.0
#> Upper-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Sutter Bypass                    0.0       0.0       0.0       0.0       0.0
#> Bear River                       0.0       0.0       0.0       0.0       0.0
#> Feather River                    0.0       0.0       0.0       0.0       0.0
#> Yuba River                       0.0       0.0       0.0       0.0       0.0
#> Lower-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Yolo Bypass                      0.0       0.0       0.0       0.0       0.0
#> American River                   0.0       0.0       0.0       0.0       0.0
#> Lower Sacramento River           0.0       0.0       0.0       0.0       0.0
#> Calaveras River                  0.0       0.0       0.0       0.0       0.0
#> Cosumnes River                   0.0       0.0       0.0       0.0       0.0
#> Mokelumne River                  0.0       0.0       0.0       0.0       0.0
#> Merced River                     0.0       0.0       0.0       0.0       0.0
#> Stanislaus River                 0.0       0.0       0.0       0.0       0.0
#> Tuolumne River                   0.0       0.0       0.0       0.0       0.0
#> San Joaquin River                0.0       0.0       0.0       0.0       0.0
#>                                   11        12        13        14        15
#> Upper Sacramento River     5149254.6 5149317.4 5149317.4 5314921.0 5332976.0
#> Antelope Creek                   0.0       0.0       0.0       0.0       0.0
#> Battle Creek                   502.4     502.4     502.4     533.8     533.8
#> Bear Creek                       0.0       0.0       0.0       0.0       0.0
#> Big Chico Creek                  0.0       0.0       0.0       0.0       0.0
#> Butte Creek                      0.0       0.0       0.0       0.0       0.0
#> Clear Creek                      0.0       0.0       0.0       0.0       0.0
#> Cottonwood Creek                 0.0       0.0       0.0       0.0       0.0
#> Cow Creek                        0.0       0.0       0.0       0.0       0.0
#> Deer Creek                       0.0       0.0       0.0       0.0       0.0
#> Elder Creek                      0.0       0.0       0.0       0.0       0.0
#> Mill Creek                       0.0       0.0       0.0       0.0       0.0
#> Paynes Creek                     0.0       0.0       0.0       0.0       0.0
#> Stony Creek                      0.0       0.0       0.0       0.0       0.0
#> Thomes Creek                     0.0       0.0       0.0       0.0       0.0
#> Upper-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Sutter Bypass                    0.0       0.0       0.0       0.0       0.0
#> Bear River                       0.0       0.0       0.0       0.0       0.0
#> Feather River                    0.0       0.0       0.0       0.0       0.0
#> Yuba River                       0.0       0.0       0.0       0.0       0.0
#> Lower-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Yolo Bypass                      0.0       0.0       0.0       0.0       0.0
#> American River                   0.0       0.0       0.0       0.0       0.0
#> Lower Sacramento River           0.0       0.0       0.0       0.0       0.0
#> Calaveras River                  0.0       0.0       0.0       0.0       0.0
#> Cosumnes River                   0.0       0.0       0.0       0.0       0.0
#> Mokelumne River                  0.0       0.0       0.0       0.0       0.0
#> Merced River                     0.0       0.0       0.0       0.0       0.0
#> Stanislaus River                 0.0       0.0       0.0       0.0       0.0
#> Tuolumne River                   0.0       0.0       0.0       0.0       0.0
#> San Joaquin River                0.0       0.0       0.0       0.0       0.0
#>                                   16        17        18         19         20
#> Upper Sacramento River     5686163.2 7962286.4 8733627.4 10061878.8 10521009.6
#> Antelope Creek                   0.0       0.0       0.0        0.0        0.0
#> Battle Creek                   596.6  285363.2  291015.2   291046.6   291140.8
#> Bear Creek                       0.0       0.0       0.0        0.0        0.0
#> Big Chico Creek                  0.0       0.0       0.0        0.0        0.0
#> Butte Creek                      0.0       0.0       0.0        0.0        0.0
#> Clear Creek                      0.0       0.0       0.0        0.0        0.0
#> Cottonwood Creek                 0.0       0.0       0.0        0.0        0.0
#> Cow Creek                        0.0       0.0       0.0        0.0        0.0
#> Deer Creek                       0.0       0.0       0.0        0.0        0.0
#> Elder Creek                      0.0       0.0       0.0        0.0        0.0
#> Mill Creek                       0.0       0.0       0.0        0.0        0.0
#> Paynes Creek                     0.0       0.0       0.0        0.0        0.0
#> Stony Creek                      0.0       0.0       0.0        0.0        0.0
#> Thomes Creek                     0.0       0.0       0.0        0.0        0.0
#> Upper-mid Sacramento River       0.0       0.0       0.0        0.0        0.0
#> Sutter Bypass                    0.0       0.0       0.0        0.0        0.0
#> Bear River                       0.0       0.0       0.0        0.0        0.0
#> Feather River                    0.0       0.0       0.0        0.0        0.0
#> Yuba River                       0.0       0.0       0.0        0.0        0.0
#> Lower-mid Sacramento River       0.0       0.0       0.0        0.0        0.0
#> Yolo Bypass                      0.0       0.0       0.0        0.0        0.0
#> American River                   0.0       0.0       0.0        0.0        0.0
#> Lower Sacramento River           0.0       0.0       0.0        0.0        0.0
#> Calaveras River                  0.0       0.0       0.0        0.0        0.0
#> Cosumnes River                   0.0       0.0       0.0        0.0        0.0
#> Mokelumne River                  0.0       0.0       0.0        0.0        0.0
#> Merced River                     0.0       0.0       0.0        0.0        0.0
#> Stanislaus River                 0.0       0.0       0.0        0.0        0.0
#> Tuolumne River                   0.0       0.0       0.0        0.0        0.0
#> San Joaquin River                0.0       0.0       0.0        0.0        0.0
#> 
#> $proportion_natural
#>                                      1           2           3           4
#> Upper Sacramento River     0.004595783 0.005968303 0.002432077 0.001944417
#> Antelope Creek                      NA          NA          NA          NA
#> Battle Creek                        NA          NA 1.000000000          NA
#> Bear Creek                          NA          NA          NA          NA
#> Big Chico Creek                     NA          NA          NA          NA
#> Butte Creek                         NA          NA          NA          NA
#> Clear Creek                         NA          NA          NA          NA
#> Cottonwood Creek                    NA          NA          NA          NA
#> Cow Creek                           NA          NA          NA          NA
#> Deer Creek                          NA          NA          NA          NA
#> Elder Creek                         NA          NA          NA          NA
#> Mill Creek                          NA          NA          NA          NA
#> Paynes Creek                        NA          NA          NA          NA
#> Stony Creek                         NA          NA          NA          NA
#> Thomes Creek                        NA          NA          NA          NA
#> Upper-mid Sacramento River          NA          NA          NA          NA
#> Sutter Bypass                       NA          NA          NA          NA
#> Bear River                          NA          NA          NA          NA
#> Feather River                       NA          NA          NA          NA
#> Yuba River                          NA          NA          NA          NA
#> Lower-mid Sacramento River          NA          NA          NA          NA
#> Yolo Bypass                         NA          NA          NA          NA
#> American River                      NA          NA          NA          NA
#> Lower Sacramento River              NA          NA          NA          NA
#> Calaveras River                     NA          NA          NA          NA
#> Cosumnes River                      NA          NA          NA          NA
#> Mokelumne River                     NA          NA          NA          NA
#> Merced River                        NA          NA          NA          NA
#> Stanislaus River                    NA          NA          NA          NA
#> Tuolumne River                      NA          NA          NA          NA
#> San Joaquin River                   NA          NA          NA          NA
#>                                      5          6          7          8
#> Upper Sacramento River     0.001676091 0.01473561 0.01993131 0.01029592
#> Antelope Creek                      NA         NA         NA         NA
#> Battle Creek                        NA         NA 1.00000000         NA
#> Bear Creek                          NA         NA         NA         NA
#> Big Chico Creek                     NA         NA         NA         NA
#> Butte Creek                         NA         NA         NA         NA
#> Clear Creek                         NA         NA         NA         NA
#> Cottonwood Creek                    NA         NA         NA         NA
#> Cow Creek                           NA         NA         NA         NA
#> Deer Creek                          NA         NA         NA         NA
#> Elder Creek                         NA         NA         NA         NA
#> Mill Creek                          NA         NA         NA         NA
#> Paynes Creek                        NA         NA         NA         NA
#> Stony Creek                         NA         NA         NA         NA
#> Thomes Creek                        NA         NA         NA         NA
#> Upper-mid Sacramento River          NA         NA         NA         NA
#> Sutter Bypass                       NA         NA         NA         NA
#> Bear River                          NA         NA         NA         NA
#> Feather River                       NA         NA         NA         NA
#> Yuba River                          NA         NA         NA         NA
#> Lower-mid Sacramento River          NA         NA         NA         NA
#> Yolo Bypass                         NA         NA         NA         NA
#> American River                      NA         NA         NA         NA
#> Lower Sacramento River              NA         NA         NA         NA
#> Calaveras River                     NA         NA         NA         NA
#> Cosumnes River                      NA         NA         NA         NA
#> Mokelumne River                     NA         NA         NA         NA
#> Merced River                        NA         NA         NA         NA
#> Stanislaus River                    NA         NA         NA         NA
#> Tuolumne River                      NA         NA         NA         NA
#> San Joaquin River                   NA         NA         NA         NA
#>                                      9          10           11           12
#> Upper Sacramento River     0.008980708 0.001267446 0.0001805059 0.0002225943
#> Antelope Creek                      NA          NA           NA           NA
#> Battle Creek                        NA          NA           NA           NA
#> Bear Creek                          NA          NA           NA           NA
#> Big Chico Creek                     NA          NA           NA           NA
#> Butte Creek                         NA          NA           NA           NA
#> Clear Creek                         NA          NA           NA           NA
#> Cottonwood Creek                    NA          NA           NA           NA
#> Cow Creek                           NA          NA           NA           NA
#> Deer Creek                          NA          NA           NA           NA
#> Elder Creek                         NA          NA           NA           NA
#> Mill Creek                          NA          NA           NA           NA
#> Paynes Creek                        NA          NA           NA           NA
#> Stony Creek                         NA          NA           NA           NA
#> Thomes Creek                        NA          NA           NA           NA
#> Upper-mid Sacramento River          NA          NA           NA           NA
#> Sutter Bypass                       NA          NA           NA           NA
#> Bear River                          NA          NA           NA           NA
#> Feather River                       NA          NA           NA           NA
#> Yuba River                          NA          NA           NA           NA
#> Lower-mid Sacramento River          NA          NA           NA           NA
#> Yolo Bypass                         NA          NA           NA           NA
#> American River                      NA          NA           NA           NA
#> Lower Sacramento River              NA          NA           NA           NA
#> Calaveras River                     NA          NA           NA           NA
#> Cosumnes River                      NA          NA           NA           NA
#> Mokelumne River                     NA          NA           NA           NA
#> Merced River                        NA          NA           NA           NA
#> Stanislaus River                    NA          NA           NA           NA
#> Tuolumne River                      NA          NA           NA           NA
#> San Joaquin River                   NA          NA           NA           NA
#>                                      13           14           15           16
#> Upper Sacramento River     3.396555e-05 3.843926e-05 1.251067e-05 0.0007681317
#> Antelope Creek                       NA           NA           NA           NA
#> Battle Creek                         NA           NA           NA           NA
#> Bear Creek                           NA           NA           NA           NA
#> Big Chico Creek                      NA           NA           NA           NA
#> Butte Creek                          NA           NA           NA           NA
#> Clear Creek                          NA           NA           NA           NA
#> Cottonwood Creek                     NA           NA           NA           NA
#> Cow Creek                            NA           NA           NA           NA
#> Deer Creek                           NA           NA           NA           NA
#> Elder Creek                          NA           NA           NA           NA
#> Mill Creek                           NA           NA           NA           NA
#> Paynes Creek                         NA           NA           NA           NA
#> Stony Creek                          NA           NA           NA           NA
#> Thomes Creek                         NA           NA           NA           NA
#> Upper-mid Sacramento River           NA           NA           NA           NA
#> Sutter Bypass                        NA           NA           NA           NA
#> Bear River                           NA           NA           NA           NA
#> Feather River                        NA           NA           NA           NA
#> Yuba River                           NA           NA           NA           NA
#> Lower-mid Sacramento River           NA           NA           NA           NA
#> Yolo Bypass                          NA           NA           NA           NA
#> American River                       NA           NA           NA           NA
#> Lower Sacramento River               NA           NA           NA           NA
#> Calaveras River                      NA           NA           NA           NA
#> Cosumnes River                       NA           NA           NA           NA
#> Mokelumne River                      NA           NA           NA           NA
#> Merced River                         NA           NA           NA           NA
#> Stanislaus River                     NA           NA           NA           NA
#> Tuolumne River                       NA           NA           NA           NA
#> San Joaquin River                    NA           NA           NA           NA
#>                                      17           18          19          20
#> Upper Sacramento River     0.0002287494 0.0008022997 0.003810468 0.008966877
#> Antelope Creek                       NA           NA          NA          NA
#> Battle Creek                         NA           NA 1.000000000 1.000000000
#> Bear Creek                           NA           NA          NA          NA
#> Big Chico Creek                      NA           NA          NA          NA
#> Butte Creek                          NA           NA          NA          NA
#> Clear Creek                          NA           NA          NA          NA
#> Cottonwood Creek                     NA           NA          NA          NA
#> Cow Creek                            NA           NA          NA          NA
#> Deer Creek                           NA           NA          NA          NA
#> Elder Creek                          NA           NA          NA          NA
#> Mill Creek                           NA           NA          NA          NA
#> Paynes Creek                         NA           NA          NA          NA
#> Stony Creek                          NA           NA          NA          NA
#> Thomes Creek                         NA           NA          NA          NA
#> Upper-mid Sacramento River           NA           NA          NA          NA
#> Sutter Bypass                        NA           NA          NA          NA
#> Bear River                           NA           NA          NA          NA
#> Feather River                        NA           NA          NA          NA
#> Yuba River                           NA           NA          NA          NA
#> Lower-mid Sacramento River           NA           NA          NA          NA
#> Yolo Bypass                          NA           NA          NA          NA
#> American River                       NA           NA          NA          NA
#> Lower Sacramento River               NA           NA          NA          NA
#> Calaveras River                      NA           NA          NA          NA
#> Cosumnes River                       NA           NA          NA          NA
#> Mokelumne River                      NA           NA          NA          NA
#> Merced River                         NA           NA          NA          NA
#> Stanislaus River                     NA           NA          NA          NA
#> Tuolumne River                       NA           NA          NA          NA
#> San Joaquin River                    NA           NA          NA          NA
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    0    0    0
#>  [2,]    0    0    0    0
#>  [3,]    0    0    0    0
#>  [4,]    0    0    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    0    0    0
#>  [7,]    0    0    0    0
#>  [8,]    0    0    0    0
#>  [9,]    0    0    0    0
#> [10,]    0    0    0    0
#> [11,]    0    0    0    0
#> [12,]    0    0    0    0
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    0
#> [16,]    0    0    0    0
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    0    0    0    0
#>