Winter Run Chinook Model

Winter Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

winter_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = winterRunDSM::params,
  stochastic = FALSE
)

Arguments

scenario
Model inputs, can be modified to test management actions

mode
The mode to run model in. Can be "seed", "simulate", "calibrate"

seeds
The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params
Parameters for model and submodels. Defaults to fallRunDSM::params.

stochastic
TRUE FALSE value indicating if model should be run stochastically. Defaults to FALSE.

Source

IP-117068

Examples

winter_run_seeds <- winterRunDSM::winter_run_model(mode = "seed")
winterRunDSM::winter_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = winter_run_seeds)

#> $spawners
#>                              1   2   3   4   5   6   7   8   9  10  11  12  13
#> Upper Sacramento River     878 766 562 478 524 554 538 478 442 442 442 442 438
#> Antelope Creek               0   0   0   0   0   0   0   0   0   0   0   0   0
#> Battle Creek                72  70  24   0   0   0   0   0   0   0   0   0   0
#> Bear Creek                   0   0   0   0   0   0   0   0   0   0   0   0   0
#> Big Chico Creek              0   0   0   0   0   0   0   0   0   0   0   0   0
#> Butte Creek                  0   0   0   0   0   0   0   0   0   0   0   0   0
#> Clear Creek                  0   0   0   0   0   0   0   0   0   0   0   0   0
#> Cottonwood Creek             0   0   0   0   0   0   0   0   0   0   0   0   0
#> Cow Creek                    0   0   0   0   0   0   0   0   0   0   0   0   0
#> Deer Creek                   0   0   0   0   0   0   0   0   0   0   0   0   0
#> Elder Creek                  0   0   0   0   0   0   0   0   0   0   0   0   0
#> Mill Creek                   0   0   0   0   0   0   0   0   0   0   0   0   0
#> Paynes Creek                 0   0   0   0   0   0   0   0   0   0   0   0   0
#> Stony Creek                  0   0   0   0   0   0   0   0   0   0   0   0   0
#> Thomes Creek                 0   0   0   0   0   0   0   0   0   0   0   0   0
#> Upper-mid Sacramento River   0   0   0   0   0   0   0   0   0   0   0   0   0
#> Sutter Bypass                0   0   0   0   0   0   0   0   0   0   0   0   0
#> Bear River                   0   0   0   0   0   0   0   0   0   0   0   0   0
#> Feather River                0   0   0   0   0   0   0   0   0   0   0   0   0
#> Yuba River                   0   0   0   0   0   0   0   0   0   0   0   0   0
#> Lower-mid Sacramento River   0   0   0   0   0   0   0   0   0   0   0   0   0
#> Yolo Bypass                  0   0   0   0   0   0   0   0   0   0   0   0   0
#> American River               0   0   0   0   0   0   0   0   0   0   0   0   0
#> Lower Sacramento River       0   0   0   0   0   0   0   0   0   0   0   0   0
#> Calaveras River              0   0   0   0   0   0   0   0   0   0   0   0   0
#> Cosumnes River               0   0   0   0   0   0   0   0   0   0   0   0   0
#> Mokelumne River              0   0   0   0   0   0   0   0   0   0   0   0   0
#> Merced River                 0   0   0   0   0   0   0   0   0   0   0   0   0
#> Stanislaus River             0   0   0   0   0   0   0   0   0   0   0   0   0
#> Tuolumne River               0   0   0   0   0   0   0   0   0   0   0   0   0
#> San Joaquin River            0   0   0   0   0   0   0   0   0   0   0   0   0
#>                             14  15  16  17  18  19  20
#> Upper Sacramento River     436 438 438 440 448 502 562
#> Antelope Creek               0   0   0   0   0   0   0
#> Battle Creek                 0   0   0   0   0   0   0
#> Bear Creek                   0   0   0   0   0   0   0
#> Big Chico Creek              0   0   0   0   0   0   0
#> Butte Creek                  0   0   0   0   0   0   0
#> Clear Creek                  0   0   0   0   0   0   0
#> Cottonwood Creek             0   0   0   0   0   0   0
#> Cow Creek                    0   0   0   0   0   0   0
#> Deer Creek                   0   0   0   0   0   0   0
#> Elder Creek                  0   0   0   0   0   0   0
#> Mill Creek                   0   0   0   0   0   0   0
#> Paynes Creek                 0   0   0   0   0   0   0
#> Stony Creek                  0   0   0   0   0   0   0
#> Thomes Creek                 0   0   0   0   0   0   0
#> Upper-mid Sacramento River   0   0   0   0   0   0   0
#> Sutter Bypass                0   0   0   0   0   0   0
#> Bear River                   0   0   0   0   0   0   0
#> Feather River                0   0   0   0   0   0   0
#> Yuba River                   0   0   0   0   0   0   0
#> Lower-mid Sacramento River   0   0   0   0   0   0   0
#> Yolo Bypass                  0   0   0   0   0   0   0
#> American River               0   0   0   0   0   0   0
#> Lower Sacramento River       0   0   0   0   0   0   0
#> Calaveras River              0   0   0   0   0   0   0
#> Cosumnes River               0   0   0   0   0   0   0
#> Mokelumne River              0   0   0   0   0   0   0
#> Merced River                 0   0   0   0   0   0   0
#> Stanislaus River             0   0   0   0   0   0   0
#> Tuolumne River               0   0   0   0   0   0   0
#> San Joaquin River            0   0   0   0   0   0   0
#> 
#> $juvenile_biomass
#>                                 1       2       3        4        5        6
#> Upper Sacramento River     3736.6 42861.0 66756.4 105629.6 108487.0 114390.2
#> Antelope Creek                0.0     0.0     0.0      0.0      0.0      0.0
#> Battle Creek                 62.8   345.4   533.8    533.8    533.8    533.8
#> Bear Creek                    0.0     0.0     0.0      0.0      0.0      0.0
#> Big Chico Creek               0.0     0.0     0.0      0.0      0.0      0.0
#> Butte Creek                   0.0     0.0     0.0      0.0      0.0      0.0
#> Clear Creek                   0.0     0.0     0.0      0.0      0.0      0.0
#> Cottonwood Creek              0.0     0.0     0.0      0.0      0.0      0.0
#> Cow Creek                     0.0     0.0     0.0      0.0      0.0      0.0
#> Deer Creek                    0.0     0.0     0.0      0.0      0.0      0.0
#> Elder Creek                   0.0     0.0     0.0      0.0      0.0      0.0
#> Mill Creek                    0.0     0.0     0.0      0.0      0.0      0.0
#> Paynes Creek                  0.0     0.0     0.0      0.0      0.0      0.0
#> Stony Creek                   0.0     0.0     0.0      0.0      0.0      0.0
#> Thomes Creek                  0.0     0.0     0.0      0.0      0.0      0.0
#> Upper-mid Sacramento River    0.0     0.0     0.0      0.0      0.0      0.0
#> Sutter Bypass                 0.0     0.0     0.0      0.0      0.0      0.0
#> Bear River                    0.0     0.0     0.0      0.0      0.0      0.0
#> Feather River                 0.0     0.0     0.0      0.0      0.0      0.0
#> Yuba River                    0.0     0.0     0.0      0.0      0.0      0.0
#> Lower-mid Sacramento River    0.0     0.0     0.0      0.0      0.0      0.0
#> Yolo Bypass                   0.0     0.0     0.0      0.0      0.0      0.0
#> American River                0.0     0.0     0.0      0.0      0.0      0.0
#> Lower Sacramento River        0.0     0.0     0.0      0.0      0.0      0.0
#> Calaveras River               0.0     0.0     0.0      0.0      0.0      0.0
#> Cosumnes River                0.0     0.0     0.0      0.0      0.0      0.0
#> Mokelumne River               0.0     0.0     0.0      0.0      0.0      0.0
#> Merced River                  0.0     0.0     0.0      0.0      0.0      0.0
#> Stanislaus River              0.0     0.0     0.0      0.0      0.0      0.0
#> Tuolumne River                0.0     0.0     0.0      0.0      0.0      0.0
#> San Joaquin River             0.0     0.0     0.0      0.0      0.0      0.0
#>                                   7        8        9       10       11
#> Upper Sacramento River     116525.4 118378.0 121392.4 122491.4 122774.0
#> Antelope Creek                  0.0      0.0      0.0      0.0      0.0
#> Battle Creek                  533.8    533.8    533.8    533.8    533.8
#> Bear Creek                      0.0      0.0      0.0      0.0      0.0
#> Big Chico Creek                 0.0      0.0      0.0      0.0      0.0
#> Butte Creek                     0.0      0.0      0.0      0.0      0.0
#> Clear Creek                     0.0      0.0      0.0      0.0      0.0
#> Cottonwood Creek                0.0      0.0      0.0      0.0      0.0
#> Cow Creek                       0.0      0.0      0.0      0.0      0.0
#> Deer Creek                      0.0      0.0      0.0      0.0      0.0
#> Elder Creek                     0.0      0.0      0.0      0.0      0.0
#> Mill Creek                      0.0      0.0      0.0      0.0      0.0
#> Paynes Creek                    0.0      0.0      0.0      0.0      0.0
#> Stony Creek                     0.0      0.0      0.0      0.0      0.0
#> Thomes Creek                    0.0      0.0      0.0      0.0      0.0
#> Upper-mid Sacramento River      0.0      0.0      0.0      0.0      0.0
#> Sutter Bypass                   0.0      0.0      0.0      0.0      0.0
#> Bear River                      0.0      0.0      0.0      0.0      0.0
#> Feather River                   0.0      0.0      0.0      0.0      0.0
#> Yuba River                      0.0      0.0      0.0      0.0      0.0
#> Lower-mid Sacramento River      0.0      0.0      0.0      0.0      0.0
#> Yolo Bypass                     0.0      0.0      0.0      0.0      0.0
#> American River                  0.0      0.0      0.0      0.0      0.0
#> Lower Sacramento River          0.0      0.0      0.0      0.0      0.0
#> Calaveras River                 0.0      0.0      0.0      0.0      0.0
#> Cosumnes River                  0.0      0.0      0.0      0.0      0.0
#> Mokelumne River                 0.0      0.0      0.0      0.0      0.0
#> Merced River                    0.0      0.0      0.0      0.0      0.0
#> Stanislaus River                0.0      0.0      0.0      0.0      0.0
#> Tuolumne River                  0.0      0.0      0.0      0.0      0.0
#> San Joaquin River               0.0      0.0      0.0      0.0      0.0
#>                                  12       13       14       15       16
#> Upper Sacramento River     127170.0 127609.6 128991.2 132413.8 158004.8
#> Antelope Creek                  0.0      0.0      0.0      0.0      0.0
#> Battle Creek                  533.8    533.8    533.8    533.8    533.8
#> Bear Creek                      0.0      0.0      0.0      0.0      0.0
#> Big Chico Creek                 0.0      0.0      0.0      0.0      0.0
#> Butte Creek                     0.0      0.0      0.0      0.0      0.0
#> Clear Creek                     0.0      0.0      0.0      0.0      0.0
#> Cottonwood Creek                0.0      0.0      0.0      0.0      0.0
#> Cow Creek                       0.0      0.0      0.0      0.0      0.0
#> Deer Creek                      0.0      0.0      0.0      0.0      0.0
#> Elder Creek                     0.0      0.0      0.0      0.0      0.0
#> Mill Creek                      0.0      0.0      0.0      0.0      0.0
#> Paynes Creek                    0.0      0.0      0.0      0.0      0.0
#> Stony Creek                     0.0      0.0      0.0      0.0      0.0
#> Thomes Creek                    0.0      0.0      0.0      0.0      0.0
#> Upper-mid Sacramento River      0.0      0.0      0.0      0.0      0.0
#> Sutter Bypass                   0.0      0.0      0.0      0.0      0.0
#> Bear River                      0.0      0.0      0.0      0.0      0.0
#> Feather River                   0.0      0.0      0.0      0.0      0.0
#> Yuba River                      0.0      0.0      0.0      0.0      0.0
#> Lower-mid Sacramento River      0.0      0.0      0.0      0.0      0.0
#> Yolo Bypass                     0.0      0.0      0.0      0.0      0.0
#> American River                  0.0      0.0      0.0      0.0      0.0
#> Lower Sacramento River          0.0      0.0      0.0      0.0      0.0
#> Calaveras River                 0.0      0.0      0.0      0.0      0.0
#> Cosumnes River                  0.0      0.0      0.0      0.0      0.0
#> Mokelumne River                 0.0      0.0      0.0      0.0      0.0
#> Merced River                    0.0      0.0      0.0      0.0      0.0
#> Stanislaus River                0.0      0.0      0.0      0.0      0.0
#> Tuolumne River                  0.0      0.0      0.0      0.0      0.0
#> San Joaquin River               0.0      0.0      0.0      0.0      0.0
#>                                  17       18       19       20
#> Upper Sacramento River     204131.4 222814.4 289822.0 319871.8
#> Antelope Creek                  0.0      0.0      0.0      0.0
#> Battle Creek                  533.8    533.8    533.8    533.8
#> Bear Creek                      0.0      0.0      0.0      0.0
#> Big Chico Creek                 0.0      0.0      0.0      0.0
#> Butte Creek                     0.0      0.0      0.0      0.0
#> Clear Creek                     0.0      0.0      0.0      0.0
#> Cottonwood Creek                0.0      0.0      0.0      0.0
#> Cow Creek                       0.0      0.0      0.0      0.0
#> Deer Creek                      0.0      0.0      0.0      0.0
#> Elder Creek                     0.0      0.0      0.0      0.0
#> Mill Creek                      0.0      0.0      0.0      0.0
#> Paynes Creek                    0.0      0.0      0.0      0.0
#> Stony Creek                     0.0      0.0      0.0      0.0
#> Thomes Creek                    0.0      0.0      0.0      0.0
#> Upper-mid Sacramento River      0.0      0.0      0.0      0.0
#> Sutter Bypass                   0.0      0.0      0.0      0.0
#> Bear River                      0.0      0.0      0.0      0.0
#> Feather River                   0.0      0.0      0.0      0.0
#> Yuba River                      0.0      0.0      0.0      0.0
#> Lower-mid Sacramento River      0.0      0.0      0.0      0.0
#> Yolo Bypass                     0.0      0.0      0.0      0.0
#> American River                  0.0      0.0      0.0      0.0
#> Lower Sacramento River          0.0      0.0      0.0      0.0
#> Calaveras River                 0.0      0.0      0.0      0.0
#> Cosumnes River                  0.0      0.0      0.0      0.0
#> Mokelumne River                 0.0      0.0      0.0      0.0
#> Merced River                    0.0      0.0      0.0      0.0
#> Stanislaus River                0.0      0.0      0.0      0.0
#> Tuolumne River                  0.0      0.0      0.0      0.0
#> San Joaquin River               0.0      0.0      0.0      0.0
#> 
#> $proportion_natural
#>                                    1         2         3          4         5
#> Upper Sacramento River     0.5034169 0.4308094 0.2241993 0.08786611 0.1679389
#> Antelope Creek             0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Battle Creek               1.0000000 1.0000000 1.0000000 0.00000000 0.0000000
#> Bear Creek                 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Big Chico Creek            0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Butte Creek                0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Clear Creek                0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Cottonwood Creek           0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Cow Creek                  0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Deer Creek                 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Elder Creek                0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Mill Creek                 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Paynes Creek               0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Stony Creek                0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Thomes Creek               0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Sutter Bypass              0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Bear River                 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Feather River              0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Yuba River                 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Yolo Bypass                0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> American River             0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Lower Sacramento River     0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Calaveras River            0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Cosumnes River             0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Mokelumne River            0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Merced River               0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Stanislaus River           0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> Tuolumne River             0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#> San Joaquin River          0.0000000 0.0000000 0.0000000 0.00000000 0.0000000
#>                                    6         7          8          9         10
#> Upper Sacramento River     0.2129964 0.1895911 0.08786611 0.01357466 0.01357466
#> Antelope Creek             0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Battle Creek               0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Bear Creek                 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Big Chico Creek            0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Butte Creek                0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Clear Creek                0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Cottonwood Creek           0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Cow Creek                  0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Deer Creek                 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Elder Creek                0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Mill Creek                 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Paynes Creek               0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Stony Creek                0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Thomes Creek               0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Sutter Bypass              0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Bear River                 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Feather River              0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Yuba River                 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Yolo Bypass                0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> American River             0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Lower Sacramento River     0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Calaveras River            0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Cosumnes River             0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Mokelumne River            0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Merced River               0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Stanislaus River           0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> Tuolumne River             0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#> San Joaquin River          0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>                                    11         12         13 14         15
#> Upper Sacramento River     0.01357466 0.01357466 0.00456621  0 0.00456621
#> Antelope Creek             0.00000000 0.00000000 0.00000000  0 0.00000000
#> Battle Creek               0.00000000 0.00000000 0.00000000  0 0.00000000
#> Bear Creek                 0.00000000 0.00000000 0.00000000  0 0.00000000
#> Big Chico Creek            0.00000000 0.00000000 0.00000000  0 0.00000000
#> Butte Creek                0.00000000 0.00000000 0.00000000  0 0.00000000
#> Clear Creek                0.00000000 0.00000000 0.00000000  0 0.00000000
#> Cottonwood Creek           0.00000000 0.00000000 0.00000000  0 0.00000000
#> Cow Creek                  0.00000000 0.00000000 0.00000000  0 0.00000000
#> Deer Creek                 0.00000000 0.00000000 0.00000000  0 0.00000000
#> Elder Creek                0.00000000 0.00000000 0.00000000  0 0.00000000
#> Mill Creek                 0.00000000 0.00000000 0.00000000  0 0.00000000
#> Paynes Creek               0.00000000 0.00000000 0.00000000  0 0.00000000
#> Stony Creek                0.00000000 0.00000000 0.00000000  0 0.00000000
#> Thomes Creek               0.00000000 0.00000000 0.00000000  0 0.00000000
#> Upper-mid Sacramento River 0.00000000 0.00000000 0.00000000  0 0.00000000
#> Sutter Bypass              0.00000000 0.00000000 0.00000000  0 0.00000000
#> Bear River                 0.00000000 0.00000000 0.00000000  0 0.00000000
#> Feather River              0.00000000 0.00000000 0.00000000  0 0.00000000
#> Yuba River                 0.00000000 0.00000000 0.00000000  0 0.00000000
#> Lower-mid Sacramento River 0.00000000 0.00000000 0.00000000  0 0.00000000
#> Yolo Bypass                0.00000000 0.00000000 0.00000000  0 0.00000000
#> American River             0.00000000 0.00000000 0.00000000  0 0.00000000
#> Lower Sacramento River     0.00000000 0.00000000 0.00000000  0 0.00000000
#> Calaveras River            0.00000000 0.00000000 0.00000000  0 0.00000000
#> Cosumnes River             0.00000000 0.00000000 0.00000000  0 0.00000000
#> Mokelumne River            0.00000000 0.00000000 0.00000000  0 0.00000000
#> Merced River               0.00000000 0.00000000 0.00000000  0 0.00000000
#> Stanislaus River           0.00000000 0.00000000 0.00000000  0 0.00000000
#> Tuolumne River             0.00000000 0.00000000 0.00000000  0 0.00000000
#> San Joaquin River          0.00000000 0.00000000 0.00000000  0 0.00000000
#>                                    16          17         18        19
#> Upper Sacramento River     0.00456621 0.009090909 0.02678571 0.1314741
#> Antelope Creek             0.00000000 0.000000000 0.00000000 0.0000000
#> Battle Creek               0.00000000 0.000000000 0.00000000 0.0000000
#> Bear Creek                 0.00000000 0.000000000 0.00000000 0.0000000
#> Big Chico Creek            0.00000000 0.000000000 0.00000000 0.0000000
#> Butte Creek                0.00000000 0.000000000 0.00000000 0.0000000
#> Clear Creek                0.00000000 0.000000000 0.00000000 0.0000000
#> Cottonwood Creek           0.00000000 0.000000000 0.00000000 0.0000000
#> Cow Creek                  0.00000000 0.000000000 0.00000000 0.0000000
#> Deer Creek                 0.00000000 0.000000000 0.00000000 0.0000000
#> Elder Creek                0.00000000 0.000000000 0.00000000 0.0000000
#> Mill Creek                 0.00000000 0.000000000 0.00000000 0.0000000
#> Paynes Creek               0.00000000 0.000000000 0.00000000 0.0000000
#> Stony Creek                0.00000000 0.000000000 0.00000000 0.0000000
#> Thomes Creek               0.00000000 0.000000000 0.00000000 0.0000000
#> Upper-mid Sacramento River 0.00000000 0.000000000 0.00000000 0.0000000
#> Sutter Bypass              0.00000000 0.000000000 0.00000000 0.0000000
#> Bear River                 0.00000000 0.000000000 0.00000000 0.0000000
#> Feather River              0.00000000 0.000000000 0.00000000 0.0000000
#> Yuba River                 0.00000000 0.000000000 0.00000000 0.0000000
#> Lower-mid Sacramento River 0.00000000 0.000000000 0.00000000 0.0000000
#> Yolo Bypass                0.00000000 0.000000000 0.00000000 0.0000000
#> American River             0.00000000 0.000000000 0.00000000 0.0000000
#> Lower Sacramento River     0.00000000 0.000000000 0.00000000 0.0000000
#> Calaveras River            0.00000000 0.000000000 0.00000000 0.0000000
#> Cosumnes River             0.00000000 0.000000000 0.00000000 0.0000000
#> Mokelumne River            0.00000000 0.000000000 0.00000000 0.0000000
#> Merced River               0.00000000 0.000000000 0.00000000 0.0000000
#> Stanislaus River           0.00000000 0.000000000 0.00000000 0.0000000
#> Tuolumne River             0.00000000 0.000000000 0.00000000 0.0000000
#> San Joaquin River          0.00000000 0.000000000 0.00000000 0.0000000
#>                                   20
#> Upper Sacramento River     0.2241993
#> Antelope Creek             0.0000000
#> Battle Creek               0.0000000
#> Bear Creek                 0.0000000
#> Big Chico Creek            0.0000000
#> Butte Creek                0.0000000
#> Clear Creek                0.0000000
#> Cottonwood Creek           0.0000000
#> Cow Creek                  0.0000000
#> Deer Creek                 0.0000000
#> Elder Creek                0.0000000
#> Mill Creek                 0.0000000
#> Paynes Creek               0.0000000
#> Stony Creek                0.0000000
#> Thomes Creek               0.0000000
#> Upper-mid Sacramento River 0.0000000
#> Sutter Bypass              0.0000000
#> Bear River                 0.0000000
#> Feather River              0.0000000
#> Yuba River                 0.0000000
#> Lower-mid Sacramento River 0.0000000
#> Yolo Bypass                0.0000000
#> American River             0.0000000
#> Lower Sacramento River     0.0000000
#> Calaveras River            0.0000000
#> Cosumnes River             0.0000000
#> Mokelumne River            0.0000000
#> Merced River               0.0000000
#> Stanislaus River           0.0000000
#> Tuolumne River             0.0000000
#> San Joaquin River          0.0000000
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    0    0    0
#>  [2,]    0    0    0    0
#>  [3,]    0    0    0    0
#>  [4,]    0    0    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    0    0    0
#>  [7,]    0    0    0    0
#>  [8,]    0    0    0    0
#>  [9,]    0    0    0    0
#> [10,]    0    0    0    0
#> [11,]    0    0    0    0
#> [12,]    0    0    0    0
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    0
#> [16,]    0    0    0    0
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    0    0    0    0
#>