Winter Run Chinook Model

Winter Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

winter_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = winterRunDSM::params,
  stochastic = FALSE
)

Source

IP-117068

Arguments

scenario

Model inputs, can be modified to test management actions

mode

The mode to run model in. Can be "seed", "simulate", "calibrate"

seeds

The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params

Parameters for model and submodels. Defaults to fallRunDSM::params.

stochastic

TRUE FALSE value indicating if model should be run stochastically. Defaults to FALSE.

Examples

winter_run_seeds <- winterRunDSM::winter_run_model(mode = "seed")
winterRunDSM::winter_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = winter_run_seeds)
#> $spawners
#>                               1    2    3    4    5    6    7    8   9  10  11
#> Upper Sacramento River     4042 3634 1736 1886 3704 3752 2908 1734 950 660 456
#> Antelope Creek                0    0    0    0    0    0    0    0   0   0   0
#> Battle Creek               3780 3594 1398  702 2474 3554 2330  734 278 156  78
#> Bear Creek                    0    0    0    0    0    0    0    0   0   0   0
#> Big Chico Creek               0    0    0    0    0    0    0    0   0   0   0
#> Butte Creek                   0    0    0    0    0    0    0    0   0   0   0
#> Clear Creek                   0    0    0    0    0    0    0    0   0   0   0
#> Cottonwood Creek              0    0    0    0    0    0    0    0   0   0   0
#> Cow Creek                     0    0    0    0    0    0    0    0   0   0   0
#> Deer Creek                    0    0    0    0    0    0    0    0   0   0   0
#> Elder Creek                   0    0    0    0    0    0    0    0   0   0   0
#> Mill Creek                    0    0    0    0    0    0    0    0   0   0   0
#> Paynes Creek                  0    0    0    0    0    0    0    0   0   0   0
#> Stony Creek                   0    0    0    0    0    0    0    0   0   0   0
#> Thomes Creek                  0    0    0    0    0    0    0    0   0   0   0
#> Upper-mid Sacramento River    0    0    0    0    0    0    0    0   0   0   0
#> Sutter Bypass                 0    0    0    0    0    0    0    0   0   0   0
#> Bear River                    0    0    0    0    0    0    0    0   0   0   0
#> Feather River                 0    0    0    0    0    0    0    0   0   0   0
#> Yuba River                    0    0    0    0    0    0    0    0   0   0   0
#> Lower-mid Sacramento River    0    0    0    0    0    0    0    0   0   0   0
#> Yolo Bypass                   0    0    0    0    0    0    0    0   0   0   0
#> American River                0    0    0    0    0    0    0    0   0   0   0
#> Lower Sacramento River        0    0    0    0    0    0    0    0   0   0   0
#> Calaveras River               0    0    0    0    0    0    0    0   0   0   0
#> Cosumnes River                0    0    0    0    0    0    0    0   0   0   0
#> Mokelumne River               0    0    0    0    0    0    0    0   0   0   0
#> Merced River                  0    0    0    0    0    0    0    0   0   0   0
#> Stanislaus River              0    0    0    0    0    0    0    0   0   0   0
#> Tuolumne River                0    0    0    0    0    0    0    0   0   0   0
#> San Joaquin River             0    0    0    0    0    0    0    0   0   0   0
#>                             12  13  14  15  16  17  18  19  20
#> Upper Sacramento River     492 422 358 372 384 408 456 616 798
#> Antelope Creek               0   0   0   0   0   0   0   0   0
#> Battle Creek                96  48   4   8  12   6   6   6   8
#> Bear Creek                   0   0   0   0   0   0   0   0   0
#> Big Chico Creek              0   0   0   0   0   0   0   0   0
#> Butte Creek                  0   0   0   0   0   0   0   0   0
#> Clear Creek                  0   0   0   0   0   0   0   0   0
#> Cottonwood Creek             0   0   0   0   0   0   0   0   0
#> Cow Creek                    0   0   0   0   0   0   0   0   0
#> Deer Creek                   0   0   0   0   0   0   0   0   0
#> Elder Creek                  0   0   0   0   0   0   0   0   0
#> Mill Creek                   0   0   0   0   0   0   0   0   0
#> Paynes Creek                 0   0   0   0   0   0   0   0   0
#> Stony Creek                  0   0   0   0   0   0   0   0   0
#> Thomes Creek                 0   0   0   0   0   0   0   0   0
#> Upper-mid Sacramento River   0   0   0   0   0   0   0   0   0
#> Sutter Bypass                0   0   0   0   0   0   0   0   0
#> Bear River                   0   0   0   0   0   0   0   0   0
#> Feather River                0   0   0   0   0   0   0   0   0
#> Yuba River                   0   0   0   0   0   0   0   0   0
#> Lower-mid Sacramento River   0   0   0   0   0   0   0   0   0
#> Yolo Bypass                  0   0   0   0   0   0   0   0   0
#> American River               0   0   0   0   0   0   0   0   0
#> Lower Sacramento River       0   0   0   0   0   0   0   0   0
#> Calaveras River              0   0   0   0   0   0   0   0   0
#> Cosumnes River               0   0   0   0   0   0   0   0   0
#> Mokelumne River              0   0   0   0   0   0   0   0   0
#> Merced River                 0   0   0   0   0   0   0   0   0
#> Stanislaus River             0   0   0   0   0   0   0   0   0
#> Tuolumne River               0   0   0   0   0   0   0   0   0
#> San Joaquin River            0   0   0   0   0   0   0   0   0
#> 
#> $juvenile_biomass
#>                                  1         2         3        4       5
#> Upper Sacramento River     44772.2 1372119.4  606874.5 770753.0 38567.5
#> Antelope Creek                 0.0       0.0       0.0      0.0     0.0
#> Battle Creek               13403.2  420559.7 1000561.2 302350.1 10353.1
#> Bear Creek                     0.0       0.0       0.0      0.0     0.0
#> Big Chico Creek                0.0       0.0       0.0      0.0     0.0
#> Butte Creek                    0.0       0.0       0.0      0.0     0.0
#> Clear Creek                    0.0       0.0       0.0      0.0     0.0
#> Cottonwood Creek               0.0       0.0       0.0      0.0     0.0
#> Cow Creek                      0.0       0.0       0.0      0.0     0.0
#> Deer Creek                     0.0       0.0       0.0      0.0     0.0
#> Elder Creek                    0.0       0.0       0.0      0.0     0.0
#> Mill Creek                     0.0       0.0       0.0      0.0     0.0
#> Paynes Creek                   0.0       0.0       0.0      0.0     0.0
#> Stony Creek                    0.0       0.0       0.0      0.0     0.0
#> Thomes Creek                   0.0       0.0       0.0      0.0     0.0
#> Upper-mid Sacramento River     0.0       0.0       0.0      0.0     0.0
#> Sutter Bypass                  0.0       0.0       0.0      0.0     0.0
#> Bear River                     0.0       0.0       0.0      0.0     0.0
#> Feather River                  0.0       0.0       0.0      0.0     0.0
#> Yuba River                     0.0       0.0       0.0      0.0     0.0
#> Lower-mid Sacramento River     0.0       0.0       0.0      0.0     0.0
#> Yolo Bypass                    0.0       0.0       0.0      0.0     0.0
#> American River                 0.0       0.0       0.0      0.0     0.0
#> Lower Sacramento River         0.0       0.0       0.0      0.0     0.0
#> Calaveras River                0.0       0.0       0.0      0.0     0.0
#> Cosumnes River                 0.0       0.0       0.0      0.0     0.0
#> Mokelumne River                0.0       0.0       0.0      0.0     0.0
#> Merced River                   0.0       0.0       0.0      0.0     0.0
#> Stanislaus River               0.0       0.0       0.0      0.0     0.0
#> Tuolumne River                 0.0       0.0       0.0      0.0     0.0
#> San Joaquin River              0.0       0.0       0.0      0.0     0.0
#>                                   6       7       8       9     10     11
#> Upper Sacramento River     986378.1 23613.1 16903.1 36266.8 8144.4 1189.1
#> Antelope Creek                  0.0     0.0     0.0     0.0    0.0    0.0
#> Battle Creek               302543.5  9431.3 10801.5 16000.5 2145.2  354.5
#> Bear Creek                      0.0     0.0     0.0     0.0    0.0    0.0
#> Big Chico Creek                 0.0     0.0     0.0     0.0    0.0    0.0
#> Butte Creek                     0.0     0.0     0.0     0.0    0.0    0.0
#> Clear Creek                     0.0     0.0     0.0     0.0    0.0    0.0
#> Cottonwood Creek                0.0     0.0     0.0     0.0    0.0    0.0
#> Cow Creek                       0.0     0.0     0.0     0.0    0.0    0.0
#> Deer Creek                      0.0     0.0     0.0     0.0    0.0    0.0
#> Elder Creek                     0.0     0.0     0.0     0.0    0.0    0.0
#> Mill Creek                      0.0     0.0     0.0     0.0    0.0    0.0
#> Paynes Creek                    0.0     0.0     0.0     0.0    0.0    0.0
#> Stony Creek                     0.0     0.0     0.0     0.0    0.0    0.0
#> Thomes Creek                    0.0     0.0     0.0     0.0    0.0    0.0
#> Upper-mid Sacramento River      0.0     0.0     0.0     0.0    0.0    0.0
#> Sutter Bypass                   0.0     0.0     0.0     0.0    0.0    0.0
#> Bear River                      0.0     0.0     0.0     0.0    0.0    0.0
#> Feather River                   0.0     0.0     0.0     0.0    0.0    0.0
#> Yuba River                      0.0     0.0     0.0     0.0    0.0    0.0
#> Lower-mid Sacramento River      0.0     0.0     0.0     0.0    0.0    0.0
#> Yolo Bypass                     0.0     0.0     0.0     0.0    0.0    0.0
#> American River                  0.0     0.0     0.0     0.0    0.0    0.0
#> Lower Sacramento River          0.0     0.0     0.0     0.0    0.0    0.0
#> Calaveras River                 0.0     0.0     0.0     0.0    0.0    0.0
#> Cosumnes River                  0.0     0.0     0.0     0.0    0.0    0.0
#> Mokelumne River                 0.0     0.0     0.0     0.0    0.0    0.0
#> Merced River                    0.0     0.0     0.0     0.0    0.0    0.0
#> Stanislaus River                0.0     0.0     0.0     0.0    0.0    0.0
#> Tuolumne River                  0.0     0.0     0.0     0.0    0.0    0.0
#> San Joaquin River               0.0     0.0     0.0     0.0    0.0    0.0
#>                                 12      13     14      15      16       17
#> Upper Sacramento River     17132.1 62408.6 3140.9 94414.3 85339.7 190015.5
#> Antelope Creek                 0.0     0.0    0.0     0.0     0.0      0.0
#> Battle Creek                4100.2 11593.4   31.4  2626.2  2931.1   2763.2
#> Bear Creek                     0.0     0.0    0.0     0.0     0.0      0.0
#> Big Chico Creek                0.0     0.0    0.0     0.0     0.0      0.0
#> Butte Creek                    0.0     0.0    0.0     0.0     0.0      0.0
#> Clear Creek                    0.0     0.0    0.0     0.0     0.0      0.0
#> Cottonwood Creek               0.0     0.0    0.0     0.0     0.0      0.0
#> Cow Creek                      0.0     0.0    0.0     0.0     0.0      0.0
#> Deer Creek                     0.0     0.0    0.0     0.0     0.0      0.0
#> Elder Creek                    0.0     0.0    0.0     0.0     0.0      0.0
#> Mill Creek                     0.0     0.0    0.0     0.0     0.0      0.0
#> Paynes Creek                   0.0     0.0    0.0     0.0     0.0      0.0
#> Stony Creek                    0.0     0.0    0.0     0.0     0.0      0.0
#> Thomes Creek                   0.0     0.0    0.0     0.0     0.0      0.0
#> Upper-mid Sacramento River     0.0     0.0    0.0     0.0     0.0      0.0
#> Sutter Bypass                  0.0     0.0    0.0     0.0     0.0      0.0
#> Bear River                     0.0     0.0    0.0     0.0     0.0      0.0
#> Feather River                  0.0     0.0    0.0     0.0     0.0      0.0
#> Yuba River                     0.0     0.0    0.0     0.0     0.0      0.0
#> Lower-mid Sacramento River     0.0     0.0    0.0     0.0     0.0      0.0
#> Yolo Bypass                    0.0     0.0    0.0     0.0     0.0      0.0
#> American River                 0.0     0.0    0.0     0.0     0.0      0.0
#> Lower Sacramento River         0.0     0.0    0.0     0.0     0.0      0.0
#> Calaveras River                0.0     0.0    0.0     0.0     0.0      0.0
#> Cosumnes River                 0.0     0.0    0.0     0.0     0.0      0.0
#> Mokelumne River                0.0     0.0    0.0     0.0     0.0      0.0
#> Merced River                   0.0     0.0    0.0     0.0     0.0      0.0
#> Stanislaus River               0.0     0.0    0.0     0.0     0.0      0.0
#> Tuolumne River                 0.0     0.0    0.0     0.0     0.0      0.0
#> San Joaquin River              0.0     0.0    0.0     0.0     0.0      0.0
#>                                  18       19       20
#> Upper Sacramento River     132858.8 235689.3 223291.3
#> Antelope Creek                  0.0      0.0      0.0
#> Battle Creek                 1800.7   2292.2   1821.2
#> Bear Creek                      0.0      0.0      0.0
#> Big Chico Creek                 0.0      0.0      0.0
#> Butte Creek                     0.0      0.0      0.0
#> Clear Creek                     0.0      0.0      0.0
#> Cottonwood Creek                0.0      0.0      0.0
#> Cow Creek                       0.0      0.0      0.0
#> Deer Creek                      0.0      0.0      0.0
#> Elder Creek                     0.0      0.0      0.0
#> Mill Creek                      0.0      0.0      0.0
#> Paynes Creek                    0.0      0.0      0.0
#> Stony Creek                     0.0      0.0      0.0
#> Thomes Creek                    0.0      0.0      0.0
#> Upper-mid Sacramento River      0.0      0.0      0.0
#> Sutter Bypass                   0.0      0.0      0.0
#> Bear River                      0.0      0.0      0.0
#> Feather River                   0.0      0.0      0.0
#> Yuba River                      0.0      0.0      0.0
#> Lower-mid Sacramento River      0.0      0.0      0.0
#> Yolo Bypass                     0.0      0.0      0.0
#> American River                  0.0      0.0      0.0
#> Lower Sacramento River          0.0      0.0      0.0
#> Calaveras River                 0.0      0.0      0.0
#> Cosumnes River                  0.0      0.0      0.0
#> Mokelumne River                 0.0      0.0      0.0
#> Merced River                    0.0      0.0      0.0
#> Stanislaus River                0.0      0.0      0.0
#> Tuolumne River                  0.0      0.0      0.0
#> San Joaquin River               0.0      0.0      0.0
#> 
#> $proportion_natural
#>                                    1         2         3         4         5
#> Upper Sacramento River     0.9148936 0.9053385 0.8018433 0.8176034 0.9071274
#> Antelope Creek             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Battle Creek               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Bear Creek                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Big Chico Creek            0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Butte Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Clear Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cottonwood Creek           0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cow Creek                  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Deer Creek                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Elder Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Mill Creek                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Paynes Creek               0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Stony Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Thomes Creek               0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sutter Bypass              0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Bear River                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Feather River              0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Yuba River                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Yolo Bypass                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> American River             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Lower Sacramento River     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Calaveras River            0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cosumnes River             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Mokelumne River            0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Merced River               0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Stanislaus River           0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Tuolumne River             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> San Joaquin River          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                                    6         7         8         9        10
#> Upper Sacramento River     0.9083156 0.8817056 0.8016148 0.6378947 0.4787879
#> Antelope Creek             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Battle Creek               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Bear Creek                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Big Chico Creek            0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Butte Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Clear Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cottonwood Creek           0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cow Creek                  0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Deer Creek                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Elder Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Mill Creek                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Paynes Creek               0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Stony Creek                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Thomes Creek               0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sutter Bypass              0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Bear River                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Feather River              0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Yuba River                 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Yolo Bypass                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> American River             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Lower Sacramento River     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Calaveras River            0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cosumnes River             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Mokelumne River            0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Merced River               0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Stanislaus River           0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Tuolumne River             0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> San Joaquin River          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                                  11       12        13         14         15
#> Upper Sacramento River     0.245614 0.300813 0.1848341 0.03910615 0.07526882
#> Antelope Creek             0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Battle Creek               1.000000 1.000000 1.0000000 1.00000000 1.00000000
#> Bear Creek                 0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Big Chico Creek            0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Butte Creek                0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Clear Creek                0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Cottonwood Creek           0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Cow Creek                  0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Deer Creek                 0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Elder Creek                0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Mill Creek                 0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Paynes Creek               0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Stony Creek                0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Thomes Creek               0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Upper-mid Sacramento River 0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Sutter Bypass              0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Bear River                 0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Feather River              0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Yuba River                 0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Lower-mid Sacramento River 0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Yolo Bypass                0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> American River             0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Lower Sacramento River     0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Calaveras River            0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Cosumnes River             0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Mokelumne River            0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Merced River               0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Stanislaus River           0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> Tuolumne River             0.000000 0.000000 0.0000000 0.00000000 0.00000000
#> San Joaquin River          0.000000 0.000000 0.0000000 0.00000000 0.00000000
#>                                   16        17       18        19        20
#> Upper Sacramento River     0.1041667 0.1568627 0.245614 0.4415584 0.5689223
#> Antelope Creek             0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Battle Creek               1.0000000 1.0000000 1.000000 1.0000000 1.0000000
#> Bear Creek                 0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Big Chico Creek            0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Butte Creek                0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Clear Creek                0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Cottonwood Creek           0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Cow Creek                  0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Deer Creek                 0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Elder Creek                0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Mill Creek                 0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Paynes Creek               0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Stony Creek                0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Thomes Creek               0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Sutter Bypass              0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Bear River                 0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Feather River              0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Yuba River                 0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Yolo Bypass                0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> American River             0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Lower Sacramento River     0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Calaveras River            0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Cosumnes River             0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Mokelumne River            0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Merced River               0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Stanislaus River           0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> Tuolumne River             0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> San Joaquin River          0.0000000 0.0000000 0.000000 0.0000000 0.0000000
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    2    0    0
#>  [2,]    0    1    0    0
#>  [3,]    0    0    0    0
#>  [4,]    0    0    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    1    0    0
#>  [7,]    0    0    0    0
#>  [8,]    0    0    0    0
#>  [9,]    0    0    0    0
#> [10,]    0    0    0    0
#> [11,]    0    0    0    0
#> [12,]    0    0    0    0
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    0
#> [16,]    0    0    0    0
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    0    0    0    0
#>